Extract index functions from SearchService
The lock mechanism is left in SearchService
to cope with the transition period.
After the update, it is unnecessary
because the framework manages locks.

The present condition is only extraction of processing.
There are many improvements in this class after the update.
